fix - change type of playerId parameter

+++ src/main/java/xyz/riocode/scoutpro/service/PlayerService.java
@@ -14,5 +14,5 @@ public interface PlayerService {
     Player getById(Long id);
     List<Player> getByName(String name);
     Player update(Player player);
-    void delete(int playerId);
+    void delete(Long playerId);
 }
